apps: enforce building golang-pie-httpserver as PIE supported by OSv
Newer Golang toolchain, like 1.15.8, creates an executable
with build option '-buildmode=pie' which crashes on OSv
as described by the issue #1141.

This patch changes the makefile to use the build option
'-buildmode=pie -ldflags "-linkmode external"' to enforce
building PIE that can receive argc and argv in a way
that is supported by OSv at the moment.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>