apps: add nginx from host example
It looks like nginx has 'daemon off;' option that allows
us to run it in a mode that does not require any fork.

This patch adds an example that pulls nginx binary along with
any dependencies from Linux host. The configuration has been
tested with version 1.18.0 and 1.20.0 with binaries taken from
Ubuntu and Fedora hosts.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>