osv-memcached: update makefile to build with modern OSv
Couple of years ago we eliminated most of the externals
that used to contain old boost and gcc headers which we now
use from host.

This patch applies a git diff to fix the makefile in the
osv-memcached project upstream. In essence it changes
the makefile to include the modules' common.gmk that provides
correct includes path.

Another positive side effect of this patch that it also
fixes the OSv issue #992 and osv-memcached runs correctly
as described in this Wiki page -
https://github.com/cloudius-systems/osv/wiki/osv-memcached.

Fixes #https://github.com/cloudius-systems/osv/issues/922

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>